Patrick,

I have merged your diffs with my own and have verified that the cumulative edits still build & run OK on my carbon version. I have uploaded my current build directory to http://www.opencpn.joat.co.uk/OpenCPN-2.3.1-Source.zip and have included the output of diffs against the original. (See diff_scr.txt, diff_inc.txt & diff_plg.txt in the root directory)

We need to establish a regular merge process to keep in step with each other. The obvious method to set up a git or subversion repository of our own pending David Register's return. Do you have access to any such resources?

What I see when running your cocoa version:
1. S52 object query work OK, unlike mine.
2. Tides & Currents don't work, the same as mine.
3. Trying to display raster charts causes an explosion of assert failures.
4. The lower chart indicator bar keeps momentarily changing size, causing the rest of the display to jump up and down by a few pixels.

Malcolm

+++ src/myiso8211/ddfrecord.cpp
@@ -282,11 +282,11 @@ int DDFRecord::Write()
 
     memset( szLeader, ' ', nLeaderSize );
 
-    sprintf( szLeader+0, "%05d", nDataSize + nLeaderSize );
+    sprintf( szLeader+0, "%05d", (int)(nDataSize + nLeaderSize) );
     szLeader[5] = ' ';
     szLeader[6] = 'D';
     
-    sprintf( szLeader + 12, "%05d", nFieldOffset + nLeaderSize );
+    sprintf( szLeader + 12, "%05d", (int)(nFieldOffset + nLeaderSize) );
     szLeader[17] = ' ';
 
     szLeader[20] = (char) ('0' + _sizeFieldLength);
